! admin fix + проблем при филтрацията
Променях малко по back-end-а, за да направя в админ панела да се трият продуктите по product number, а не по product id.  Трябва още да се направи в shop на 255 ред да се взима product.tag.gender за да се реализира филтрация на продуктите
Също трябва да се направи и бек енда на количката

+++ src/main/java/bg/conquerors/wardrobe/service/impl/AdminServiceImpl.java
@@ -12,6 +12,7 @@ import bg.conquerors.wardrobe.service.AdminService;
 import org.springframework.stereotype.Service;
 
 import java.time.temporal.TemporalAdjuster;
+import java.util.List;
 
 @Service
 public class AdminServiceImpl implements AdminService {
@@ -68,12 +69,21 @@ public class AdminServiceImpl implements AdminService {
     }
 
     @Override
-    public void deleteProduct(Long id) {
-        var product = productRepository.findAllById(id);
-        Tag tag = product.getTag();
-        productRepository.delete(product);
-        tagRepository.delete(tag);
+    public void deleteProduct(String productNumber) {
+        List<Product> products = productRepository.findAllByProductNumber(productNumber);
+        if (!products.isEmpty()) {
+            for (Product product : products) {
+                 Tag tag = product.getTag();
+                productRepository.delete(product);
+                 if (tag != null) {
+                     tagRepository.delete(tag);
+                 }
+            }
+        } else {
+            throw new IllegalArgumentException("No products found with product number: " + productNumber);
+        }
     }
+
 
     private Product getNewProduct(AddProductDTO addProductDTO) {
         Product createProduct = new Product();
